+++ ChangeLog.md
@@ -1,5 +1,16 @@
 # OldAS ChangeLog
 
+## Unrealsed
+
+**Released: WiP**
+
+- Added `ArticleIDs.full_ids`.
+  ([#10](https://github.com/davep/oldas/pull/10))
+- Added `ArticleIDs.mark_read`.
+  ([#10](https://github.com/davep/oldas/pull/10))
+- Added `ArticleIDs.mark_unread`.
+  ([#10](https://github.com/davep/oldas/pull/10))
+
 ## v0.3.1
 
 **Released: 2026-01-07**